+++ NEWS.md
@@ -1,4 +1,4 @@
-# review development
+# review 3.8.0
 
 ## New features and changes